# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ScripturesApp API",
    "description": "Backend for ScripturesApp Native App",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health Check",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/search": {
      "get": {
        "summary": "Search Verses",
        "operationId": "search_verses_api_v1_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bible reference (e.g. 'Gn 1:1')",
              "title": "Q"
            },
            "description": "Bible reference (e.g. 'Gn 1:1')"
          },
          {
            "name": "tr",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Translations to show (en, fr, gr, hb, ar)",
              "title": "Tr"
            },
            "description": "Translations to show (en, fr, gr, hb, ar)"
          },
          {
            "name": "v",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Primary version (N1904, LXX, BHSA)",
              "default": "N1904",
              "title": "V"
            },
            "description": "Primary version (N1904, LXX, BHSA)"
          },
          {
            "name": "bible",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "French version (tob, bj)",
              "title": "Bible"
            },
            "description": "French version (tob, bj)"
          },
          {
            "name": "crossref",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Show cross references",
              "default": false,
              "title": "Crossref"
            },
            "description": "Show cross references"
          },
          {
            "name": "crossref_full",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Display cross-references with text",
              "default": false,
              "title": "Crossref Full"
            },
            "description": "Display cross-references with text"
          },
          {
            "name": "crossref_source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter cross-references by source",
              "title": "Crossref Source"
            },
            "description": "Filter cross-references by source"
          },
          {
            "name": "crossref_max",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Max cross-references kept per verse for margin display",
              "default": 3,
              "title": "Crossref Max"
            },
            "description": "Max cross-references kept per verse for margin display"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerseResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/find": {
      "get": {
        "summary": "Find Words",
        "description": "Find occurrences of a word or expression.\nDetects script (Greek vs Latin) to determine search mode.",
        "operationId": "find_words_api_v1_find_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Search query (Greek word or French expression)",
              "title": "Q"
            },
            "description": "Search query (Greek word or French expression)"
          },
          {
            "name": "tr",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Translations to show (en, fr, gr, hb, ar)",
              "title": "Tr"
            },
            "description": "Translations to show (en, fr, gr, hb, ar)"
          },
          {
            "name": "v",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Greek Corpus (nt, lxx, all)",
              "title": "V"
            },
            "description": "Greek Corpus (nt, lxx, all)"
          },
          {
            "name": "bible",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "French version (tob, bj)",
              "title": "Bible"
            },
            "description": "French version (tob, bj)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Max results",
              "default": 20,
              "title": "Limit"
            },
            "description": "Max results"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "post": {
        "summary": "Search",
        "operationId": "search_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SearchResult"
                  },
                  "type": "array",
                  "title": "Response Search Search Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CrossReferenceRelationDTO": {
        "properties": {
          "target_ref": {
            "type": "string",
            "title": "Target Ref"
          },
          "target_ref_localized": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Ref Localized"
          },
          "rel_type": {
            "$ref": "#/components/schemas/CrossReferenceType"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "target_ref_margin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Ref Margin"
          }
        },
        "type": "object",
        "required": [
          "target_ref",
          "rel_type"
        ],
        "title": "CrossReferenceRelationDTO"
      },
      "CrossReferenceType": {
        "type": "string",
        "enum": [
          "parallel",
          "quotation",
          "allusion",
          "other"
        ],
        "title": "CrossReferenceType"
      },
      "FindResponseDTO": {
        "properties": {
          "lemma": {
            "type": "string",
            "title": "Lemma",
            "description": "The query lemma or expression"
          },
          "original": {
            "type": "string",
            "title": "Original",
            "description": "The original query string"
          },
          "lemma_gloss": {
            "type": "string",
            "title": "Lemma Gloss",
            "description": "Gloss/Definition of the lemma",
            "default": ""
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of results found"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/FindResultItemDTO"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "lemma",
          "original",
          "total"
        ],
        "title": "FindResponseDTO"
      },
      "FindResultItemDTO": {
        "properties": {
          "ref": {
            "type": "string",
            "title": "Ref",
            "description": "Verse reference (e.g. 'Mt 1:1')"
          },
          "book_code": {
            "type": "string",
            "title": "Book Code",
            "description": "Standard book code"
          },
          "chapter": {
            "type": "integer",
            "title": "Chapter"
          },
          "verse": {
            "type": "integer",
            "title": "Verse"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Main text content (Greek or French)"
          },
          "translations": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Translations",
            "description": "Parallel translations"
          },
          "highlights": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Highlights",
            "description": "Words to highlight in the main text"
          }
        },
        "type": "object",
        "required": [
          "ref",
          "book_code",
          "chapter",
          "verse",
          "text"
        ],
        "title": "FindResultItemDTO"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Language": {
        "type": "string",
        "enum": [
          "en",
          "fr",
          "gr",
          "hb",
          "ar"
        ],
        "title": "Language"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerseCrossReferencesDTO": {
        "properties": {
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/CrossReferenceRelationDTO"
            },
            "type": "array",
            "title": "Relations"
          }
        },
        "type": "object",
        "title": "VerseCrossReferencesDTO"
      },
      "VerseDTO": {
        "properties": {
          "book_code": {
            "type": "string",
            "title": "Book Code"
          },
          "chapter": {
            "type": "integer",
            "title": "Chapter"
          },
          "verse": {
            "type": "integer",
            "title": "Verse"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "book_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Name"
          },
          "node": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "book_code",
          "chapter",
          "verse",
          "text",
          "language",
          "version"
        ],
        "title": "VerseDTO"
      },
      "VerseItemDTO": {
        "properties": {
          "ref": {
            "type": "string",
            "title": "Ref"
          },
          "primary": {
            "$ref": "#/components/schemas/VerseDTO"
          },
          "parallels": {
            "items": {
              "$ref": "#/components/schemas/VerseDTO"
            },
            "type": "array",
            "title": "Parallels"
          },
          "cross_references": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerseCrossReferencesDTO"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "ref",
          "primary"
        ],
        "title": "VerseItemDTO"
      },
      "VerseResponseDTO": {
        "properties": {
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "verses": {
            "items": {
              "$ref": "#/components/schemas/VerseItemDTO"
            },
            "type": "array",
            "title": "Verses"
          },
          "cross_references": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerseCrossReferencesDTO"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "reference",
          "verses"
        ],
        "title": "VerseResponseDTO"
      },
      "SearchQuery": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 10
          },
          "translation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation"
          },
          "book": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchQuery"
      },
      "SearchResult": {
        "properties": {
          "ref": {
            "type": "string",
            "title": "Ref"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "translation": {
            "type": "string",
            "title": "Translation"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "book": {
            "type": "string",
            "title": "Book"
          },
          "chapter": {
            "type": "integer",
            "title": "Chapter"
          },
          "verse": {
            "type": "integer",
            "title": "Verse"
          }
        },
        "type": "object",
        "required": [
          "ref",
          "text",
          "translation",
          "score",
          "book",
          "chapter",
          "verse"
        ],
        "title": "SearchResult"
      }
    }
  }
}